requirements.txt
asgiref==3.3.1
autopep8==1.5.6
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
